wip: mac_mapper refactor

+++ src/if_service.c
@@ -71,7 +71,7 @@ bool create_subnet_ifs(UT_array *ifinfo_array, bool ignore_error)
   }
 
   while(p = (config_ifinfo_t*) utarray_next(ifinfo_array, p)) {
-    snprintf(longip, IP_LEN,"%s/%d", p->ip_addr, get_short_subnet(p->subnet_mask));
+    snprintf(longip, IP_LEN,"%s/%d", p->ip_addr, (int)get_short_subnet(p->subnet_mask));
 
     log_trace("Creating ifname=%s ip_addr=%s brd_addr=%s subnet_mask=%s", p->ifname, p->ip_addr, p->brd_addr, p->subnet_mask);
